Add files via upload
Credit Card Number Checker

This was a small project from Codecademy. It takes a nested array of credit card numbers and returns an array of "bad numbers" (numbers that failed the Luhn algorithm test). This new array can then be run through a function that determines which credit card company each of the bad numbers corresponds to, and outputs this list into another array.